pyproject.toml
[tool.poetry]
name = "pgdictgen"
version = "0.1.0"
description = "Data dictionary generator by Damico"
authors = ["Jose Damico <jose.damico@scicrop.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.9"
reportlab = "^3.5.67"
sshtunnel = "^0.4.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
